Move betty.plugins.command to betty.commands
