+++ .coveragerc
@@ -1,6 +1,6 @@
 [run]
 omit =
-    betty/plugins/command/dev_profile_demo.py
+    betty/commands/dev_profile_demo.py
     betty/sphinx/extension/betty.py
 branch = True
 concurrency =
@@ -11,7 +11,7 @@ source = betty
 
 [report]
 omit =
-    betty/plugins/command/dev_profile_demo.py
+    betty/commands/dev_profile_demo.py
     betty/sphinx/extension/betty.py
 exclude_also =
     ^\s*pragma: no cover$
