+++ betty/commands/dev_update_translations.py
