+++ betty/commands/generate.py
