+++ betty/commands/new.py
