+++ betty/commands/new_translation.py
