+++ betty/tests/commands/test_about.py
