+++ betty/tests/commands/test_clear_caches.py
@@ -44,7 +44,7 @@ class TestClearCaches:
             legacy_cache_directory = tmp_path / "legacy-cache"
             legacy_cache_directory.mkdir()
             mocker.patch(
-                "betty.plugins.command.clear_caches._LEGACY_CACHE_DIRECTORY",
+                "betty.commands.clear_caches._LEGACY_CACHE_DIRECTORY",
                 legacy_cache_directory,
             )
             legacy_cache_item = legacy_cache_directory / "item"
