+++ betty/tests/commands/test_dev_update_translations.py
