+++ betty/tests/commands/test_generate.py
