+++ betty/tests/commands/test_new.py
