+++ pyproject.toml
@@ -114,18 +114,18 @@ betty = 'betty.console:main_standalone'
 'wiki' = 'betty.plugins.asset_directory.wiki:WIKI'
 
 [project.entry-points.'betty.command']
-'about' = 'betty.plugins.command.about:About'
-'clear-caches' = 'betty.plugins.command.clear_caches:ClearCaches'
-'config' = 'betty.plugins.command.config:Config'
-'demo' = 'betty.plugins.command.demo:Demo'
-'dev-profile-demo' = 'betty.plugins.command.dev_profile_demo:_discover'
-'dev-update-translations' = 'betty.plugins.command.dev_update_translations:_discover'
-'docs' = 'betty.plugins.command.docs:Docs'
-'generate' = 'betty.plugins.command.generate:Generate'
-'new' = 'betty.plugins.command.new:New'
-'new-translation' = 'betty.plugins.command.new_translation:NewTranslation'
-'serve' = 'betty.plugins.command.serve:Serve'
-'update-translations' = 'betty.plugins.command.update_translations:UpdateTranslations'
+'about' = 'betty.commands.about:About'
+'clear-caches' = 'betty.commands.clear_caches:ClearCaches'
+'config' = 'betty.commands.config:Config'
+'demo' = 'betty.commands.demo:Demo'
+'dev-profile-demo' = 'betty.commands.dev_profile_demo:_discover'
+'dev-update-translations' = 'betty.commands.dev_update_translations:_discover'
+'docs' = 'betty.commands.docs:Docs'
+'generate' = 'betty.commands.generate:Generate'
+'new' = 'betty.commands.new:New'
+'new-translation' = 'betty.commands.new_translation:NewTranslation'
+'serve' = 'betty.commands.serve:Serve'
+'update-translations' = 'betty.commands.update_translations:UpdateTranslations'
 
 [project.entry-points.'betty.copyright_notice']
 'project-author' = 'betty.plugins.copyright_notice.project_author:ProjectAuthor'
